// src/modules/solfin/dtos/createNewKyc.dto.ts
export class CreateNewKycSolfin {
    identificacion: string;
    tipo_id: string;
    autoridad: string;
    nombre?: string;
    primer_apellido?: string;
    segundo_apellido?: string;
    business_name?: string
    business_nickname?: string
    sexo?: number;
    direccion: string;
    telefono: string;
    movil: string;
    correo: string;
    actividadeconomica: string;
    profesion: string;
    tiporesidencia: string;
    cargo: string;
    nivelestudio: string;
    marital: string;
    id_expiration_date: string;
    fecha_nacimiento: string;
    upload_front: string;
    upload_back: string;
    nationality: number;
    id_pais: number;
    id_provincia: string;
    id_canton: string;
    id_zona: string;
    monthly_income: number;
    privacy: boolean;
}