// src/modules/solfin/interfaces/kycStatus.ts
export interface SolfinKycStatus {
    status: string;
    document: string;
    documentType: number;
    statusDescription: string;
    sinpeId: string;
    name: string;
    lastName: string;
    fullname: string;
    businessName: string;
    address: string;
    expirationDate: string;
    nationality: string;
    phone: string;
    mobile: string;
    email: string;
    picture: string;
}